Refactor consultation services for improved modularity
- Replace service implementations with corresponding interfaces for better abstraction (`IConsultationService`, `IConsultationPreparationService`, `IReservationService`).
- Remove unused and redundant service implementations (`ConsultationPreparationServiceImpl`, `ConsultationSearchServiceImpl`, `ConsultationServiceImpl`, `ReservationServiceImpl`).
- Update dependencies to use the new interfaces across the application.
- Adjust unit tests to reflect the renamed and updated classes (`ConsultationSearchService`, `ConsultationPreparationService`, etc.).
