+++ src/main/java/com/nexaworks/rafiq/service/consultation/ConsultationService.java
@@ -1,17 +1,161 @@
 package com.nexaworks.rafiq.service.consultation;
 
+import java.time.LocalDateTime;
 import java.util.UUID;
 
+import org.springframework.dao.PessimisticLockingFailureException;
+import org.springframework.messaging.simp.SimpMessagingTemplate;
+import org.springframework.retry.annotation.Backoff;
+import org.springframework.retry.annotation.Retryable;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import org.springframework.transaction.support.TransactionSynchronization;
+import org.springframework.transaction.support.TransactionSynchronizationManager;
+
+import com.nexaworks.rafiq.dto.event.ConsultationChanged;
 import com.nexaworks.rafiq.dto.request.consultation.AddConsultationRequest;
-import com.nexaworks.rafiq.entities.Consultation;
+import com.nexaworks.rafiq.entities.*;
 import com.nexaworks.rafiq.entities.enums.ConsultationStatus;
+import com.nexaworks.rafiq.exception.custom.ConsultationCanNotCancelledException;
+import com.nexaworks.rafiq.exception.custom.ConsultationCanNotEditException;
+import com.nexaworks.rafiq.exception.custom.ConsultationException;
+import com.nexaworks.rafiq.repository.ConsultationRepository;
+import com.nexaworks.rafiq.repository.DoctorRepository;
+import com.nexaworks.rafiq.service.authentication.AuthService;
+import com.nexaworks.rafiq.service.rabbit.MessageService;
 
-public interface ConsultationService {
-    Consultation add(Consultation request);
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 
-    Consultation editConsultation(AddConsultationRequest request, UUID id);
+@Service
+@RequiredArgsConstructor
+@Slf4j
+public class ConsultationService implements IConsultationService {
+    private final ConsultationRepository consultationRepository;
+    private final AuthService authService;
+    private final DoctorRepository doctorRepository;
+    private final SimpMessagingTemplate messagingTemplate;
+    private final MessageService messageService;
+    private final SimpMessagingTemplate simpMessagingTemplate;
 
-    void expire(String consultationId);
+    @Override
+    @Transactional(rollbackFor = Exception.class)
+    @Retryable(retryFor = {
+            PessimisticLockingFailureException.class}, maxAttempts = 3, backoff = @Backoff(delay = 500))
+    public Consultation add(Consultation entity) {
+        Doctor doctor = (Doctor) authService.getAuthenticateUser();
+        entity.setDoctor(doctor);
 
-    void update(UUID id, ConsultationStatus consultationStatus);
+        doctorRepository.findByIdWithLock(doctor.getId());
+
+        LocalDateTime startTime = entity.getTimeSlot().getStartTime();
+        LocalDateTime endTime = entity.getTimeSlot().getStartTime()
+                .plusMinutes(entity.getTimeSlot().getDurationMinutes());
+
+        if (consultationRepository.existsByOverlapping(startTime, endTime, doctor.getId(),
+                ConsultationStatus.CANCELLED)) {
+            throw new ConsultationCanNotCancelledException(
+                    "Consultation time slot is already booked");
+        }
+        entity.getTimeSlot().setEndTime(endTime);
+
+        Consultation consultation = consultationRepository.save(entity);
+
+        log.info("Consultation added {} by {}", consultation.getId(), doctor.getEmail());
+
+        TransactionSynchronizationManager.registerSynchronization(new TransactionSynchronization() {
+            @Override
+            public void afterCommit() {
+                messageService.publishExpirationEvent(consultation.getId(),
+                        consultation.getTimeSlot().getEndTime());
+            }
+        });
+        return consultation;
+    }
+
+    @Override
+    @Transactional(rollbackFor = Exception.class)
+    @Retryable(retryFor = {
+            PessimisticLockingFailureException.class}, maxAttempts = 3, backoff = @Backoff(delay = 500))
+    public Consultation editConsultation(AddConsultationRequest request, UUID id) {
+        UUID userId = authService.getAuthenticateUserId();
+
+        Consultation consultation = consultationRepository.findById(id)
+                .orElseThrow(() -> new ConsultationException("Consultation not found"));
+
+        validateEditability(consultation, userId);
+
+        doctorRepository.findByIdWithLock(userId);
+
+        LocalDateTime start = request.startTime();
+        LocalDateTime end = request.startTime().plusMinutes(request.duration());
+        if (consultationRepository.existsByOverlapping(start, end, userId, id,
+                ConsultationStatus.CANCELLED)) {
+            throw new ConsultationCanNotEditException("Consultation time slot is already booked");
+        }
+        consultation.getTimeSlot().setStartTime(start);
+        consultation.getTimeSlot().setEndTime(end);
+        log.info("Consultation edited {} by {}", consultation.getId(), userId);
+
+        TransactionSynchronizationManager.registerSynchronization(new TransactionSynchronization() {
+            @Override
+            public void afterCommit() {
+                messagingTemplate.convertAndSend("/topic/consultation",
+                        new ConsultationChanged(consultation.getId(), start));
+            }
+        });
+
+        return consultationRepository.save(consultation);
+    }
+
+    private static void validateEditability(Consultation consultation, UUID userId) {
+        if (consultation.getStatus().isTerminal()) {
+            throw new ConsultationCanNotEditException("Consultation is already completed");
+        }
+        if (consultation.getStatus() != ConsultationStatus.AVAILABLE) {
+            throw new ConsultationCanNotEditException("You cannot edit on booked consultation");
+        }
+        if (!consultation.getDoctor().getId().equals(userId)) {
+            throw new ConsultationException("You are not authorized to edit this consultation");
+        }
+    }
+
+    @Override
+    @Transactional
+    public void expire(String consultationId) {
+        Consultation consultation = consultationRepository
+                .findConsultationById(UUID.fromString(consultationId))
+                .orElseThrow(() -> new ConsultationException("Consultation not found"));
+        if (consultation.getStatus() == ConsultationStatus.AVAILABLE) {
+            consultation.setStatus(ConsultationStatus.EXPIRED);
+
+        } else if (consultation.getStatus() == ConsultationStatus.LIVE) {
+            consultation.setStatus(ConsultationStatus.COMPLETED);
+        }
+
+        consultationRepository.save(consultation);
+        log.info("Consultation expired {}", consultationId);
+        TransactionSynchronizationManager.registerSynchronization(new TransactionSynchronization() {
+            @Override
+            public void afterCommit() {
+                if (consultation.getStatus() == ConsultationStatus.AVAILABLE) {
+                    messageService.publishDoctorExpireNotification(
+                            consultation.getDoctor().getNotificationToken(),
+                            consultation.getTimeSlot().getStartTime());
+
+                } else if (consultation.getStatus() == ConsultationStatus.LIVE) {
+                    simpMessagingTemplate.convertAndSend("/queue/" + consultationId, 1);
+                }
+            }
+        });
+    }
+
+    @Override
+    public void update(UUID id, ConsultationStatus consultationStatus) {
+        Consultation consultation = consultationRepository.findConsultationById(id)
+                .orElseThrow(() -> new ConsultationException("Consultation not found"));
+        consultation.setStatus(consultationStatus);
+        consultationRepository.save(consultation);
+    }
+
 }
