+++ src/main/java/com/nexaworks/rafiq/service/consultation/IConsultationPreparationService.java
@@ -1,0 +1,7 @@
+package com.nexaworks.rafiq.service.consultation;
+
+import java.util.UUID;
+
+public interface IConsultationPreparationService {
+    void prepare(UUID uuid);
+}
