+++ src/main/java/com/nexaworks/rafiq/service/consultation/IReservationService.java
@@ -1,0 +1,10 @@
+package com.nexaworks.rafiq.service.consultation;
+
+import java.util.UUID;
+
+import com.nexaworks.rafiq.entities.enums.PaymentProvider;
+import com.stripe.exception.StripeException;
+
+public interface IReservationService {
+    String reserve(UUID id, PaymentProvider provider) throws StripeException;
+}
